requirements-win.txt
# Windows 自动化依赖
uiautomation
pyperclip
pywin32
Pillow
